add workflowbuilder folder package to separate, plus separated builder workbench for API development
