+++ desktopclient/qml/datapanel/modeller/OperationObject.qml
@@ -1,18 +1,40 @@
 import QtQuick 2.0
 
 BasicModellerObject {
+    id: operationObject
 
     readonly property string nameText: "Operation"
 
-   property int defaultWidth : 50
-   property int defaultHeight : 50
+    property int defaultWidth : 50
+    property int defaultHeight : 50
 
     width: defaultWidth
     height: defaultHeight
 
+    Image {
+        id: icon
+        source: "../../world.svg"
+        //source: "../../images/workflow/operation.svg"
+        width: operationObject.width
+        height: operationObject.height
+    }
+
+    function scale(factor) {
+        icon.width = operationObject.width * factor
+        icon.height = operationObject.height * factor
+    }
+
     function draw(ctx) {
         ctx.save();
         ctx.beginPath();
+
+        ctx.drawImage(icon, x-width/2, y-height/2, 1500, 1500);
+        ctx.text("Leck mich fett!!1", (x - width/2) + 20, y);
+        ctx.stroke();
+
+        ctx.scale(10,10)
+
+        /*
         if (selected) {
             ctx.lineWidth = 2;
             ctx.strokeStyle = "red";
@@ -39,6 +61,9 @@ BasicModellerObject {
             ctx.text(nameText,(x - width/2) + 20, y);
         }
         ctx.stroke();
+        */
+
+
         ctx.restore();
     }
