--- ilwiscoreui/models/workflowmodel.cpp
@@ -1,6 +1,0 @@
-#include "workflowmodel.h"
-
-WorkflowModel::WorkflowModel(QObject *parent) :
-    QObject(parent)
-{
-}
